Adapt filter-node-props to changes in org-element.
Apparently, a property-drawer is now tightly coupled to an org headline
directly above it. If that does not exist, the exact same
property-drawer becomes a drawer with name="property". Thus, the former
code, that narrowed buffer to only the org-element of
interest (`org-dp-contents') does not work anymore for property-drawers,
because these change their type due to the narrowing to drawer.

Therefore, getting the content could not be delegated anymore to
`org-dp-contents' but has to be done here, using the associated headline
and not the property-drawer as container.